Update Page.cs
Add on two methods for the program to use. Haven't had the chance to test them yet, but they should give the contents of the correct table, and give you whether or not that table entry would have the data entry you are looking for. The main method is the same. Whoever is going to make the code to traverse the hierarchy, tell me if you have any problems with the code. Should have getPhysicalAddress() return the full physical address as the end result in the coming days, with offset added to it.