Finish getPhysicalAddress
Made the method getPhysicalAddress output the hexadecimal address for the physical memory, which includes the offset. Also made the variables declared in what is now the PageTable constructor available for the entire class, not just the constructor. Not tested yet, but is error-free and theoretically sound. Any problems with the code can be directed to me.